show only photos that has been fetched in the container (gallery, certain tag, or certain album)

+++ PhotoBox/Controllers/AlbumsViewController.m
@@ -114,7 +114,7 @@
 #pragma mark - Did stuff
 
 - (void)didFetchItems {
-    int count = [self.dataSource numberOfItems];
+    NSInteger count = [self.dataSource numberOfItems];
     [self setAlbumsCount:count max:self.totalItems];
 }
